Move files to async folder
