+++ lib/src/async/capabilities.dart
