+++ lib/src/async/command_event.dart
