+++ lib/src/async/command_processor.dart
