+++ lib/src/async/options.dart
