+++ lib/src/async/window.dart
